test: Path 클래스 단위 테스트 작성중
